requirements.txt
Django == 3.1
djangorestframework
django-cors-headers
djangorestframework-jwt
flake8
black
coverage
psycopg2-binary
django-rest-passwordreset
django-extensions
pytest-django
django-filter
django-environ
Pillow
pyqrcode
pypng